+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file.
 - Fix grouping and sorting in the text version of the report/nag emails (#679)
 - Fix the update_timeout view to record more information for downtime bookkeeping
 - Fix the checks list to preserve selected filters when adding/updating checks (#684)
+- Fix duration calculation to skip "log" and "ign" events
 
 ## v2.2.1 - 2022-06-13